requirements.txt
python-telegram-bot==20.7
spotipy==2.23.0
yt-dlp==2023.12.30
python-dotenv==1.0.0
discord.py==2.3.2
PyNaCl==1.5.0
ffmpeg-python==0.2.0